Polymorphic types
The types of functions (and other values) may now have contain type
variables, for example,
  hd: 'a list -> 'a
and those type variables are resolved when the function is applied:
  hd [1,2,3]
has type "int".

We do not support polymorphic "let", e.g.
  let fun id x = x in (id 1, id true) end
or polymorphic datatype, e.g.
  datatype 'a tree ...

This is implemented by a new kind of type, ForallType. The type of
"hd", for instance, is "forall 'a. 'a list -> 'a". We print this
as "'a list -> 'a", but the "forall" section indicates which type
variables are bound.

Give type variables names 'a, 'b, etc.

In the unifier, make Variable implement Comparable, so that a is before
b is before z is before T0 is before T1 is before T10 etc. (Variables of
the form "T<integer>" are sorted after explicitly named variables, and
are compared using their integer ordinals.)

Upgrade to guava-21 so that we can use ImmutableList collector.

Add ConsList (copied from Calcite).