[MOREL-20] Add `order` clause
`order`, `where` and `group` can occur after `from`, in any order,
repeated as  many times as you like.

Unlike `group`, `order` propagates its input variables. For example,
if you have `from e in emps, d in depts`, then variables `d` and `e`
will available for use in the expressions of an `order` clause that
follows it, and they will also be available for the `group`, `where`,
`order` or `yield` that follows the `order`.

This change does not implement comparator functions mentioned in the
issue. Just `exp` and `exp desc`.