[MOREL-23] Default arguments for aggregate functions
If you omit `of`, the argument will be the variable(s) defined by
enclosing `from` clause or preceding `from` step.

Add a test to demonstrate that you can use the identity function
`fn x => x` as an aggregate function, and it produces the same result
as SQL's COLLECT, namely the collection of rows in the current group.