[MOREL-28] The `sum` aggregate function only works on `int` values
Make "sum" work for both int and float (by making it a Macro rather than
a function). Add "min" and "max".

In Environments, tidy up how we build compilation environments with
built-in values, foreign values, etc. And add MapEnvironment that
contains many bindings in a single map.

During compilation, improve how we derive names and types for the
intermediate steps in a "from" pipeline.