[MOREL-29] Overload `+` operator to allow both `int` and `real` arguments
It is tricky to overload '+' (and other arithmetic operators), because
if there are no clues in the argument types it defaults to 'int'.
Such operators now declare their 'preferred type', and if the type is
still a type variable after type resolution pass, we set the type to
the preferred type and re-try.

When generating code, such operators now use a Macro, which looks at the
reified type and generates code accordingly.

Refactor operators (~ + - * / div mod < <= > >= = <> :: ^) so that they
are handled like built-in functions.

Convert the implementation of several built-ins from Code to Applicable
(or Macro if overloaded).