[MOREL-33] Add `elem`, `notElem` operators
SQL has two `IN` operators (one for a sub-query, one for a list) and
because queries are lists in Morel, this operator is equivalent to both.

However, we cannot use `in` because that is already in the language
(inherited from Standard ML). It would be ambiguous in an expression
inside `let`. For example, `let val b = 10 in deptnos in b orelse b2 end`
is ambiguous.

We could disallow `in` in these contexts only by duplicating large parts
of the grammar to add 'in-free expressions'. Not worth it.

Here is an example:
```
from d in scott.dept
where d.deptno inside (from e in scott.emp
                       where e.job inside ["ANALYST", "PRESIDENT"]
                       yield e.deptno);
val it =
  [{deptno=10,dname="ACCOUNTING",loc="NEW YORK"},
   {deptno=20,dname="RESEARCH",loc="DALLAS"}]
  : {deptno:int, dname:string, loc:string} list
```

`notElem` is equivalent to SQL's `NOT IN`. `e notElem list` is also
equivalent to `not (e elem list)` but often requires fewer parentheses.