Deduce type of polymorphic field in tuple or record
The problem was that after applying an action during
unification (when we have deduced the type of a record
selector function) we did not check whether the variable had
already been expanded by unification.

Create a lighter weight Substitution class, for use during
the unification algorithm, whose map is not immutable and not
sorted.