AST that contains a call to "op +" is unparsed incorrectly
If the first argument to Apply is an Op (referencing the
built-in "op +") unparse is correct; but if the first
argument to Apply is an Id whose value is "op +" (that is,
an unresolved reference to an operator), then it is
incorrectly unparsed as with function rather than infix
syntax.

When unparsing calls to Apply, the solution is to convert
operators such as "op +" to built-ins in order to get
correct infix, precedence and associativity.