[MOREL-13] Garbage-collect obscured variables
When binding a variable "v" to a SubEnvironment, if the
SubEnvironment also binds "v", bind to the parent instead.
This prevents chains of environments from forming, and
allows the obscured "v" to be garbage-collected.

When binding 5 or more variables at the same time (and
therefore creating a MapEnvironment) see whether the
variables obscure ALL variables in the parent environment,
and if so, make EmptyEnvironment.INSTANCE the parent
environment.