[MOREL-45] Translate List.filter as if user had written where; and List.map to yield
Convert calls to List.filter and List.map to
Core.From expressions, so that programs written in the
functional style produce the same Calcite algebra as those
written in the relational style (using 'from').

NOTE: This change cannot handle multiple calls to `List.map`,
which corresponds to a `from` with multiple `yield` clauses.
[MOREL-52], coming shortly, will improve that particular
case.

Convert calls to List.filter and List.map to Core.From expressions