[MOREL-51] Add `Relational.exists` function, and push it down to Calcite
Also `Relational.notExists`, because the extra parentheses
in `not (exists (from ...))` are a pain.

'exists' can be used as EXISTS in SQL - to test whether a
sub-query returns any rows - and can also be used as an
aggregate function.

In future, if you use 'exists' in a Morel expression, it
will become a semi-join in Calcite's relational algebra.

Improve how correlated sub-queries are translated.