[MOREL-70] Polymorphic datatype
Allow declaration of polymorphic algebraic types, for example

  datatype 'a tree = EMPTY | NODE of 'a tree * 'a * 'a tree

and also with multiple type parameters, for example

  datatype ('a, 'b) Union = A of 'a | B of 'b

and mutually recursive, for example

  datatype 'a tree = Empty | Node of 'a * 'a forest
  and      'a forest = Nil | Cons of 'a tree * 'a forest

We already support polymorphism in built-in types, functions,
records, tuples, for example

  'a list
  'a option
  ('a * 'b) list -> 'a.

Creating self-referential datatypes, and groups of datatypes
that reference each other, is tricky. So we add `interface
Type.Key` to be an immutable blueprint for the type that we
are seeking. If it already exists, we can look it up by key;
if not, we can create it.

We also add `class TemporaryType extends DataType` to act as
a placeholder while we are wiring up a datatype (or creating
a copy of a datatype with real types sustituted for type
variables, e.g. converting `'a tree` to `int tree` and fixing
up its constructors to match). And add `class Transaction` to
remove temporary types when we're done.

Add `Codes.EMPTY_ENV`, so that we don't create a new
environment from scratch every time we execute a statement.

Fixes #70