[MOREL-72] Join
Add clauses to `from` to support inner joins. We continue to
allow comma joins, but only up until the first step (`where`,
`join`, `group`, `yield` or `order` keyword). After that,
commas would introduce ambiguity when combined with the
commas in `group` or `compute`.

We will add outer joins (`left`, `right`, `full` keywords)
in a later commit.

Unlike in SQL, the `on` clause is optional; without it,
`on true` is assumed, which makes for a cartesian join. But
it will allow us to do cartesian outer joins, which are
useful sometimes.

Joins, like all clauses in from, are left-associative. If you
want to create a bushy join, e.g.
((a join b) join (c join d)), you will need to use a nested
`from` expression.