[MOREL-94] Wordle solver
This code example solves the Wordle word puzzle. It includes
the corpus of words and solutions used by the actual web site,
and has a mechanism for choosing the optimal guess given the
information from previous guesses.

As future work, we could benchmark this solver against other
solvers, and see which takes fewest guesses on average.