[MOREL-99] Script parser hangs if it encounters invalid syntax, and ignores comments at ends of files
The script parser (used by tests that run `.sml` files, and
also by the `use` command) hangs if it encounters invalid
syntax. The cause is that the parser buffers the input that
it reads for each command, so that it can echo that input to
the output.

Previously, when the parser encountered an error, it would
try to parse again from the same point; now it skips over
that input. The parser also now aborts if the input is empty
(that is, we're not making any progress).

Lastly, if there are no errors, the parser now continues
reaching input until it reaches the end of the file.
Previously it looked for a command followed by a semicolon.
The effect was to ignore comments at the ends of files.

Fixes #99