[MOREL-103] Layered patterns ('as'), and composite 'val'
Previously we only allow identifiers on the left-hand side of
a declaration in a 'val'. With composite 'val', we allow
other patterns, such as tuples, records, and layered patterns.

Layered patterns allow you to assign a variable for both the
whole pattern and (some of) the parts. For example,

  val p as (i, b) = (1, true);

assigns i = 1, b = true and p = (1, true).

You can also use layered patterns in relational
expressions:

  from p as (i, b) in [(1, true), (2, false)];

Fixes #103