Move project to 'hydromatic' GitHub organization
The goal is to signal that contributions are welcome. It also allows
cruft (e.g. development branches and tags) to stay in the julianhyde
repo, as would be the case for any other contributor.

The home URL has changed from https://github.com/julianhyde/morel
to https://github.com/hydromatic/morel. The previous URL is still
valid, but is now a fork.

Issues, discussions and releases are now at hydromatic, and
unfortunately the URLs don't redirect; many of the development
branches and tags have been removed, and only exist in the julianhyde
fork.