[MOREL-107] ShellTest is non-deterministic
The problem seems to be that JLine has a race-condition.

The solution is for the GitHub action to set an environment
variable 'morel.ci' when running 'mvn test'; the
non-deterministic tests in ShellTest check that variable and
skip if it is set.

When run from the command-line, all tests in ShellTest will
run as normal. This ensures that they get run occasionally.
Non-determinism seems to be less of a problem from the
command-line, perhaps because there are more cores available,
and perhaps because it is easier to re-run.

Fixes #107