Satisfiability prover
Implementation is brute force. If there are N variables, the
algorithm will try each combination of assigments of {false,
true} to variables, and stop when it finds one that works.

We can improve the implementation later, if necessary. For
example, if there is a group of variables where exactly one
is always true, then we can eliminate one of the variables.
(If exactly one of A, B and C is true, replace A with "not
(B or C)".)

For the environment, use a boolean[] rather than a
Map<Variable, Boolean>, exploiting the fact that variables have
contiguous integer ids.