[MOREL-55] Analyze match coverage, detecting redundant and exhaustive matches
Allow the shell to print warnings (even multiple warnings).

Add property 'matchCoverageEnabled', to control whether
coverage is analyzed. If it is false, an expression will
never be rejected for being redundant.

Add $bool and $list internal data types. This allows true,
false, NIL and CONS to be analyzed in matches the same way as
other constructors. Unlike values of other types, we know that
constructors are a finite set. So we can deduce that

  fun f true = ...
    | f false = ...

is exhaustive. Because $bool and $list are internal, they
can't be used in programs and don't appear in metadata.

Fixes #55