Add interface Tracer, so that tests can check for several events during the lifecycle
In particular, the method `Tracer.onCore(int pass, Core.Decl e)`
is called at various points in the conversion of an `Ast.Decl`
to a `Core.Decl`, and allows a test to check on progress at
any point that it chooses. In contrast to the previous
approach, the preparation only occurs once.