[MOREL-198] Idempotent mode for test scripts
To date, scripts had .sml suffix and were accompanied by a
reference log with .sml.out suffix; with this change, scripts
may have a .smli suffix, and such scripts can have expected
output interleaved, and therefore no reference file is needed.
The expected output is on lines prefixed '> ', and is ignored
when reading commands from the file.

If the script is successful, the generated .smli file will
match the source .smli file; if the generated file contains
the correct output, you should copy the generated file, so
that it becomes the new source.

Idempotent scripts are easier to read (the expected output
is right there) and merge (you can inspect and manually merge
expected output, rather than re-generating the output and
hoping for the best).

Fixes #198