[MOREL-199] Support JDK 19 and 20
The main problem is that the algorithm to convert floating
point numbers to strings changed in JDK 19, because
https://bugs.openjdk.org/browse/JDK-4511638 has been fixed
in the JDK.

In this change, we switch to the 'correct' floating point
representation, which is the one generated by JDK 19 and up.
But how to achieve this representation in earlier JDKs? In
this commit, we look for specific values that we know to
occur in tests, and replace them with the correct values.
It's not a great solution, because as we add tests in
future, we may enounter new values that need to be masked.

In a future commit we could try to do better; say, re-create
the JDK 19 algorithm (as long as we can do it without
including a heftly library), or fix up values to match.
There is some initial work in the
https://github.com/julianhyde/morel/tree/199-jdk19 branch.

Add `class JavaVersion`, to allow switching functionality on
and off depending on the JDK version.

Add JDK 19 and 20 to CI.

Fixes hydromatic/morel#199