[MOREL-200] In the test suite, run arbitrary "lint" checks on code
Add a test class LintTest that runs arbitrary checks on the
source code. It also checks for tabs and trailing spaces, and
checks all textual source files.

The test is similar to the LintTest added in [CALCITE-5765],
and also uses the Puffin script engine, added in
[CALCITE-5764].

Upgrade Calcite to 1.36.0.

Fixes hydromatic/morel#200