Update (jdk to 21, calcite to 1.36.0, guava to 32.1.3-jre, hsqldb to 2.7.2, jline to 3.24.1, checkstyle to 10.12.4)
We also update a few Maven plugins; see diff for details.

The lowest supported JDK and Guava versions has not changed.
We now support (and test) JDK 8 through 21, and Guava 21.0
through 32.1.3-jre.

This commit adds the 'versions' plugin to the Maven build
file. It will allow us to easily update dependencies to
their latest release. To do so, run the following command:
  ./mvnw versions:update-properties