[MOREL-209] File reader, and progressive types
Adds a type-safe system to browse directories, subdirectories,
and read files as lists of records.

If Morel is started with the argument
`--directory=src/test/resources/data`, you can write

  from d in file.scott.dept;

because `data` has a subdirectory `scott` that contains a
file `dept.csv`.

This change introduces *progressive types* to solve the
problem that, when browsing a large file system, we would
have to traverse every directory, and parse every file,
in order to report the type of the `file` value. The type of
`file` is progressive; it is at first a partial record (with
missing fields shown as `, ...`), but over time, more fields
in that record are discovered. Fields are never lost, so
once a program has been typed, the type remains valid.

Add sample files under `src/test/resources/data`;
migrate `wordle.smli` to use the Wordle sample data.

In this version we support `.csv` and `.csv.gz` files;
later, we could support other file types, or just delegate to
Calcite's file adapter.

Add `class TypedValue`, to represent a value that knows its
own type, so that the value's type can change during the
session.

The new built-in `file` value is not global, but belongs to
a session, to prevent thread-safety issues, and to prevent
concurrent tests from interfering with each other. Add
`BuiltIn.sessionValue`.

During type deduction (unification) we add a label, `z$dummy`,
that distinguishes progressive record types from regular
record types. It is removed before the record type is created.

Add article based on screencast https://youtu.be/uybUjCYsBKI.

Fixes #209