[MOREL-171] Add `into` and `through` clauses to `from` expression
`into` and `through` allow you to pass the rows of a query
into a what SQL would call a "table function", conveniently
placing the table function at the end of the pipeline.

`into` must be the last step of a `from` expression, and
sends the rows of the query to a function, as an argument of
type `list`. For example,

  from e in scott.emp
    yield e.deptno
    into sum

is equivalent to

  sum (from e in scott.emp
    yield e.deptno)

`through` is similar but does not need to be the last step.
The function needs to accept an argument of type `list` and
also return a `list`. A scan commences, with the pattern `p`
in `through p in f` being successively bound to each element
of the list returned by `f`.

The following example uses a `pivot` function, that converts
x rows of y columns into y rows of x columns:

  from list in [[1,2,3,4],[5,6,7,8]]
    through list2 in pivot;
  > val it = [[1,5],[2,6],[3,7],[4,8]] : int list list

  from list in [[1,2,3,4],[5,6,7,8]]
    through list2 in pivot
    yield List.nth (list2, 0) + List.nth (list2, 1);
  > val it = [6,8,10,12] : int list

Fixes #171