Validate field references
Drop the requirement that a non-terminal `yield` step is a
record. Such a step does not produce any named fields, but
can still be used anonymously; for example, the following
query is valid:

  from e in scott.emp
    yield e.deptno
    compute sum

If the step after the `yield` attempts to reference fields by
name, no name will be valid. When

  from e in scott.emp
    yield e.deptno * 2
    yield x - 1

gives the error "unbound variable or constructor: x", the
query can be made valid by using a record to name the field:

  from e in scott.emp
    yield {x = e.deptno * 2}
    yield x - 1