Add function `Interactive.useSilently`
`useSilently file` has the same effect as `use file`, but
does not print to stdout. It is particularly suitable when a
test (`.smli` file) wants to include a preamble but does not
want to include that preamble's output in its own output.

`ScriptTest` should verify that the included file generates
the correct output (e.g. identical to the input, if it is an
`.smli` file, or identical to the `.ref` for other file
types) but it does not currently verify.

Add scott dataset to `relational.smli` by including new file
`scott.smli` using `useSilently`. After the include, there
is a variable `scott` that has fields `emp`, `dept`, `bonus`,
`salgrade`, like the JDBC data source, but stored on the
heap, and in fact overshadowing the JDBC data source.