[MOREL-231] Add `distinct` keyword as shorthand for `group` with all fields and no aggregate functions
Add distinct keyword as shorthand for group with all fields and
no aggregate functions. For example, the following are
equivalent:

  (*) Using `distinct`
  from e in scott.emp
    yield {e.job, e.deptno}
    distinct;

  (*) Using `group` with no aggregate functions
  from e in scott.emp
    yield {e.job, e.deptno}
    group job, deptno;

This is analogous to how, in SQL,

  SELECT DISTINCT job, deptno FROM emp

is equivalent to

  SELECT job, deptno FROM emp GROUP BY job, deptno

SQL has a complication: you are allowed to use DISTINCT in the
same SELECT block as a GROUP BY, and it is applied after
aggregate functions are evaluated. Thankfully Morel does not
have to worry about SELECT blocks.

The implementation adds a `Distinct` class to the AST but
desugars it to `Group` in the core algebra.

Fixes #231