[MOREL-273] Ordered and unordered queries
Inputs to queries can now be ordered collections (`list`),
unordered collections (`bag`), or a combination. The output
collection type of a query depends on the collection types
of the inputs; generally, it is ordered only if all inputs
are ordered, or if it includes an `order` step.

A query used to be `ListType`, now it must be a collection
type (`ListType` or instance of the `bag` datatype). Add the
method `Type.isCollection()`.

In `group`, `compute` and `into`, the collection passed to
an aggregate function is ordered iff the input is ordered.
This will change in #271.

Make `interface MultiType` a top-level class

Add `class EnvironmentTypeEnv`, a type environment backed by
an `Environment`.

In `TypeResolver`, make `deducePatType` return `void`.

Logged #270, #271, #272 for follow-up

Fixes #273