[MOREL-241] Universal and existential quantification (`forall` and `exists`) and implication (`implies`)
Examples:
```
  (* Do any employees have a salary under 900? *)
  exists e in emps
    where e.sal < 900.0;

  (* Do all employees have a salary greater than 1,000? *)
  forall e in emps
    require e.sal > 1000.0;

  (* Do all employees in department 10 earn over 1,000? *)
  forall e in emps
    require e.deptno = 10 implies e.sal > 1000.0;

  (* Find all departments where all employees earn more than 1,000. *)
  from d in depts
    where (forall e in emps
      require e.deptno = d.deptno
      implies e.sal > 1000.0);
```

To avoid clashing with the new `exists` keyword, we had to
quote uses of functions `List.exists` and `Vector.exists`,
and rename a couple of user-defined functions.

Rename `Relational.exists` to `nonEmpty`, and
`Relational.notExists` to `empty`. If you would write a query
using `EXISTS` in SQL, idiomatic Morel would use `exists`.
(This is a breaking change.)

`empty` and `nonEmpty` are still useful as aggregate
functions, equivalent to `count = 0` and `count > 0`.

Fixes #241