[MOREL-245] Enable google-java-format
`google-java-format` reformats code to a standard style.
It reduces merge conflicts because all PRs have the same
style.

`LintTest` (added in #200) still performs its checks;
Checkstyle also runs, but does less than before.

In the previous commit, we added a Maven plugin for
`google-java-format`, but it was disabled; in this commit,
we reformat source files to comply, and we enable the plugin.
There are no functionality changes in this commit, and all
changes to Java files are automatic.

If you have a PR that you wish to merge onto main, you should
rebase onto the previous commit, set skip=false in pom.xml,
and build. This will standardize the format of your code so
that you can more easily continue your merge.

Fixes #245