[MOREL-248] Optimize method `Static.transformEager` for `List` and `Collection` arguments
If argument is a `Collection`, we know the size in
advance; we can return an empty list of the argument is
empty otherwise pre-allocate an `ImmutableList.Builder` of
the correct size. If the argument is a `List`, we can
further optimize, skipping the builder if the list has one
element.

This change was only possible after bumping guava to 23.1,
because we need the method
`ImmutableList.builderWithExpectedSize(int)`.

Fixes #248