Bump guava from 33.4.0-jre to 33.4.5-jre, and raise minimum from 21.0 to 23.1-jre
Increase the 'typical' version in CI from 23.0 to 31.1-jre.

Version 23.1 allows us to use the method
`ImmutableList.builderWithExpectedSize(int)`.

Version 21.0 allows us to use the method
`ImmutableList.toImmutableList()` and remove
`Static.toImmutableList()`.

Add `order` clause to a few test queries, to ensure that the
ordering of results is the same in guava 31.1-jre as other
versions.