Refactor `NameGenerator`
Make `NameGenerator` global. If we are running in a shell,
the `Session` owns a generator and all commands within it
will use that generator. This ensures that internal variable
names generated for one command do not overlap internal
variable names generated for another command. This becomes
important when we generate internal names for overloaded
functions.

With a global generator, it is more likely that variables
will have non-zero ordinals, and these are unparsed with
names like `foo_2`. We therefore fix up the environment that
is used when Calcite executes Morel code fragments,
converting the variable `foo` with ordinal 2 to a variable
`foo_2` with ordinal 0. We should come up with a better
solution.

Reduce coupling to `NameGenerator` by using its `get()`
method as a `Supplier<String>` and is `inc()` function as a
`ToIntFunction<String>`.

Add '$' to generated variable names; 'v0' becomes 'v$0', etc.

Remove a copy in `EnvShuttle` that may not be necessary.