The `toString` methods of `RelList` and `Binding` should not print the contents of tables
`Binding.toString() and `RelList.toString()` are not called
in production code, only in the debugger, but even so, they
were causing the debugger to be slow and use a lot of memory.

If you really want to see the whole contents (presumably
only on small tables), call the new method
`RelList.asString()`.