[MOREL-249] Add `with` operator (functional update notation for record values)
Inspired by OCaml's `with` syntax. If `r` is a record with
type `{x: string, y: int, z: bool}` then the expression
`{r with y = 5, z = true}` has the same type as `r` and
inherits the value of field `x` from `r`. It is a convenient
syntax to modify a small number of a record's fields (or,
more precisely, to copy the record and retain the values of
all fields not explicitly changed).

Fixes #249