[MOREL-260] Add function `Sys.showAll ()`
Add function `Sys.showAll ()` (abbreviated `showAll ()`) that
shows the name and value of every property. This extends
support for session variables (show, set, unset) added in #44.

Example:
```
  - showAll ();
  val it =
    [("directory",SOME "/Users/jhyde/dev/morel.2"),("hybrid",SOME "false"),
     ("inlinePassCount",SOME "5"),("lineWidth",SOME "79"),
     ("matchCoverageEnabled",SOME "true"),("optionalInt",NONE),
     ("printDepth",SOME "5"),("printLength",SOME "12"),
     ("relationalize",SOME "false"),
     ("scriptDirectory",SOME "/Users/jhyde/dev/morel.2"),
     ("stringDepth",SOME "70")] : (string * string option) list
```

Note that property `optionalInt` (added for testing purposes)
has value `NONE`, which means that it is not set.

Refactor `misc.smli` so that fewer lines change each time
a function is added to the built-in environment.

Fixes #260