[MOREL-259] Tabular mode
Add "output" property with values "classic" and "tabular".
If the mode is "tabular" and the value is a list of records,
the result is
```
  - Sys.set ("output", "tabular");
  val it = () : unit
  - from e in scott.dept;
  deptno dname      loc
  ------ ---------- --------
      10 ACCOUNTING NEW YORK
      20 RESEARCH   DALLAS
      30 SALES      CHICAGO
      40 OPERATIONS BOSTON

  val it : {deptno:int, dname:string, loc:string} bag
```

as opposed to the classic mode:
```
  - Sys.set ("output", "tabular");
  val it = () : unit
  - from e in scott.dept;
  val it =
    [{deptno=10,dname="ACCOUNTING",loc="NEW YORK"},
     {deptno=20,dname="RESEARCH",loc="DALLAS"},
     {deptno=30,dname="SALES",loc="CHICAGO"},
     {deptno=40,dname="OPERATIONS",loc="BOSTON"}]
    : {deptno:int, dname:string, loc:string} bag
```

Currently the fields of the record must be primitive types
(not records, lists, or sum types such as option), otherwise
Morel reverts to classic mode.

In `class Prop` allow properties to have enum types, and to
be mandatory/optional.

When parsing shell output, allow treat a line containing
only '>' as an output line. Previously the line had to start
with '> '.

Fixes #259