[MOREL-262] Allow atomic `yield` steps at any point in the pipeline
Previously, Morel only allowed an atomic `yield` - a `yield`
step whose expression was not a record constructor - if it
was the last step in the query. The reason was that the
expression did not have a natural name. But it precluded
some reasonable queries, such as

```
  from i in [1, 2, 3, 4, 5, 6],
      j in [1, 2, 3, 4, 5, 6]
    yield i + j
    distinct
```

which does not need to reference any fields.

In this commit, we let you write such queries, and we derive
a name for the singleton field when possible. If the step is
`yield deptno` or `yield e.deptno`, the singleton field is
called `deptno`.

(In a future change (#265) we will add the `current` keyword
so that you can always refer to the current row. We also
need to do better at inferring field names in steps with
`with`, such as `yield {e with sal = sal * 2.0}`)

For implementation, we add `class Core.StepEnv`, so that
each step knows not only its list of bindings but also
whether its value is an atom. (An atom must have exactly one
binding, but a step with one binding is not necessarily an
atom.) We also remove `Query.implicitQueryExp`, which is no
longer necessary.

Fixes #262