[MOREL-269] Decompose the documentation for built-in functions and structures
The reference contains a table (in Markdown format) of all
built-in functions. This change moves the metadata in that
table to a few file, functions.toml, adds a test to generate
that section of the reference from the metadata and check
that it is the same.

In the future, functions will be added to the metadata first,
and we will generate more things from the metadata.

The table will also allow us to generate finer-grained
documentation, maybe one page for each structure or function.

Obsolete `class TestUtil`, moving methods to `TestUtils`.

Add `Bool.op implies` to function table.

Fixes #269