[MOREL-282] Add built-in datatype `Descending`, and method `Relational.compare`, for type-based orderings
This change allows us to create expressions and types that
express every possible ordering scheme. We add a built-in
datatype `Descending` (with a single constructor `DESC`) to
reverse the ordering of its argument type.

We also add a method `Relational.compare`, of type
`'a * 'a -> order`, which can compare a pair of values of
any type. `Relational.compare` is not a typical function.
Its implementation needs reified types.  We achieve this by
adding a macro-like facility: the new method
`Applicable.withType` is called any time a built-in is
converted to a constant.

`Descending` and `Relational.compare` will enable `order` to
have a single expression (#244), in the next commit.

Constructors of a datatype have an intrinsic order, not
ordered by name.

In TypeSystem, combine methods `bag` and `bagType`.

Refactor method `Type.copy`.

Fixes #282